Update Rest Service
1. Add service for Paid Record
2. Add time format in application.properties

+++ Rest Service/src/main/java/com/xiushui/application/entity/PaidRecord.java
@@ -28,7 +28,7 @@ public class PaidRecord
 	private long id;
 	
 	@Column(nullable = false)
-	private int userDonateId;
+	private long userDonateId;
 	
 	@Column(nullable = false)
 	private int paid;
